+++ src/spaceinvaders/model/GameObject.java
@@ -5,11 +5,11 @@ package spaceinvaders.model;
  */ 
 
 public abstract class GameObject {
-    private int sizeX;
-    private int sizeY;
+    private double sizeX;
+    private double sizeY;
 
-    private int posX;
-    private int posY;
+    private double posX;
+    private double posY;
 
 	private boolean alive;
 
@@ -24,7 +24,7 @@ public abstract class GameObject {
  	 * @param posX - position on X axis
 	 * @param posY - position on Y axis
 	 */
-    GameObject(int posX, int posY) {
+    GameObject(double posX, double posY) {
         this.posX = posX;
         this.posY = posY;
 
@@ -41,7 +41,7 @@ public abstract class GameObject {
 	 * @param sizeX - width of object
 	 * @param sizeY - height of object
 	 */
-    GameObject(int posX, int posY, int sizeX, int sizeY) {
+    GameObject(double posX, double posY, double sizeX, double sizeY) {
         this.posX = posX;
         this.posY = posY;
 
@@ -56,7 +56,7 @@ public abstract class GameObject {
 	 * Method used to move object on each game refresh
 	 * @param steps
 	 */
-	void move(int steps){
+	void move(double steps){
 		switch (direction){
 			case up:
 				posY = posY - steps;
@@ -98,14 +98,14 @@ public abstract class GameObject {
 		this.move(1);
 	}
 	/** Getter for variable direction
-	 * @return direction the object is pointing at
+	 * @return direction the object is podoubleing at
 	 */
 	public Direction getDirection() {
 		return direction;
 	}
 
 	/** Setter for variable direction
-	 * @param direction - direction the object is pointing at
+	 * @param direction - direction the object is podoubleing at
 	 */
 	void setDirection(Direction direction) {
 		this.direction = direction;
@@ -113,28 +113,28 @@ public abstract class GameObject {
 	/** Getter for position on X axis
 	 * @return position on X axis
 	 */
-    public final int getPosX()
+    public final double getPosX()
     {
         return posX;
     }
 	/** Getter for position on Y axis
 	 * @return position on Y axis
 	 */
-    public final int getPosY()
+    public final double getPosY()
     {
         return posY;
     }
 	/** Getter for width of object
 	 * @return width of object
 	 */
-    public final int getSizeX()
+    public final double getSizeX()
     {
         return sizeX;
     }
 	/** Getter for height of object
 	 * @return height of object
 	 */
-    public final int getSizeY()
+    public final double getSizeY()
     {
         return sizeY;
     }
@@ -142,7 +142,7 @@ public abstract class GameObject {
 	 * Setter for position on X axis
 	 * @param posX position on X axis
 	 */
-    public void setPosX(int posX) {
+    public void setPosX(double posX) {
         if(posX > 0)
             this.posX = posX;
     }
@@ -150,21 +150,21 @@ public abstract class GameObject {
 	 * Setter for position on Y axis
 	 * @param posY position on Y axis
 	 */
-    public void setPosY(int posY) {
+    public void setPosY(double posY) {
         if(posY > 0)
             this.posY = posY;
     }
 	/** Setter for width of object
 	 * @param sizeX width of object
 	 */
-    public void setSizeX(int sizeX) {
+    public void setSizeX(double sizeX) {
         if(sizeX > 0)
             this.sizeX = sizeX;
     }
 	/** Setter for height of object
 	 * @param sizeY height of object
 	 */
-    public void setSizeY(int sizeY) {
+    public void setSizeY(double sizeY) {
         if(sizeY > 0)
             this.sizeY = sizeY;
     }
@@ -173,7 +173,7 @@ public abstract class GameObject {
 	 * @param X how many steps in direction of X axis
 	 * @return position on X axis
 	 */
-    public int modifyPosX(int X) {
+    public double modifyPosX(double X) {
         this.posX += X;
 		return this.posX;
     }
@@ -182,7 +182,7 @@ public abstract class GameObject {
 	 * @param Y how many steps in direction of Y axis
 	 * @return position on Y axis
 	 */
-    public int modifyPosY(int Y) {
+    public double modifyPosY(double Y) {
         this.posY += Y;
 		return this.posY;
     }
@@ -191,7 +191,7 @@ public abstract class GameObject {
 	 * @param X how many units in direction of X axis
 	 * @return size on X axis or -1 if size would be less than 0
 	 */
-    public int modifySizeX(int X) {
+    public double modifySizeX(double X) {
 		if(X<this.sizeX)
 		{
 			this.sizeX += X;
@@ -206,7 +206,7 @@ public abstract class GameObject {
 	 * @param Y how many units in direction of Y axis
 	 * @return size on Y axis or -1 if size would be less than 0
 	 */
-    public int modifySizeY(int Y) {
+    public double modifySizeY(double Y) {
 		if(Y<this.sizeY) {
 			this.sizeY += Y;
 			return this.sizeY;
